processor/aggregationprocessor/factory.go
// Copyright  observIQ, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package aggregationprocessor

import (
	"context"
	"fmt"
	"time"

	"github.com/observiq/observiq-otel-collector/processor/aggregationprocessor/internal/aggregate"
	"go.opentelemetry.io/collector/component"
	"go.opentelemetry.io/collector/consumer"
	"go.opentelemetry.io/collector/processor"
)

const (
	typeStr = "aggregation"

	stability = component.StabilityLevelAlpha
)

// NewFactory creates a new ProcessorFactory with default configuration
func NewFactory() processor.Factory {
	return processor.NewFactory(
		typeStr,
		createDefaultConfig,
		processor.WithMetrics(createMetricsProcessor, stability),
	)
}

func createDefaultConfig() component.Config {
	return &Config{
		Interval: 1 * time.Minute,
		Include:  "^.*$",
		Aggregations: []AggregateConfig{
			{
				Type:       aggregate.AggregateTypeMin,
				MetricName: "$0.min",
			},
			{
				Type:       aggregate.AggregateTypeMax,
				MetricName: "$0.max",
			},
			{
				Type:       aggregate.AggregateTypeAvg,
				MetricName: "$0.avg",
			},
		},
	}
}

func createMetricsProcessor(
	ctx context.Context,
	set processor.CreateSettings,
	cfg component.Config,
	nextConsumer consumer.Metrics,
) (processor.Metrics, error) {
	oCfg, ok := cfg.(*Config)
	if !ok {
		return nil, fmt.Errorf("cannot create aggregation processor with invalid config type: %t", cfg)
	}

	sp, err := newAggregationProcessor(set.Logger, oCfg, nextConsumer)
	if err != nil {
		return nil, fmt.Errorf("failed to create aggregation processor: %w", err)
	}

	return sp, nil
}
